+++ client/public/index.html
@@ -4,16 +4,16 @@
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width,initial-scale=1.0">
-    <link rel="icon" href="<%= BASE_URL %>icon.png">
+    <link rel="icon" href="<%= BASE_URL %>images/icon.png">
     <title>XSS Catcher</title>
   </head>
   <body>
     <noscript>
-      <strong>We're sorry but <%= htmlWebpackPlugin.options.title %> doesn't work properly without JavaScript enabled. Please enable it to continue.</strong>
+      <strong>We're sorry but XSS Catcher doesn't work properly without JavaScript enabled. Please enable it to continue.</strong>
     </noscript>
     <div id="app"></div>
     <!-- built files will be auto injected -->
   </body>
-  <link rel="stylesheet" type="text/css" href="/static/bootstrap.min.css">
-  <link rel="stylesheet" type="text/css" href="/static/duotone-sea.css">
+  <link rel="stylesheet" type="text/css" href="/static/styles/bootstrap.min.css">
+  <link rel="stylesheet" type="text/css" href="/static/styles/duotone-sea.css">
 </html>
