+++ client/public/static/scripts/collector.js
