+++ client/src/components/index/viewData/DeleteData.vue
@@ -1,0 +1,48 @@
+<template>
+  <b-modal
+    ref="deleteDataModal"
+    id="delete-data-modal"
+    title="Are you sure?"
+    hide-footer
+  >
+    <b-form>
+      <b-button @click="deleteData" variant="outline-danger"
+        >Yes, delete this entry</b-button
+      >
+      <b-button
+        @click="$refs.deleteDataModal.hide()"
+        variant="outline-secondary"
+        >Cancel</b-button
+      >
+    </b-form>
+  </b-modal>
+</template>
+<script>
+import axios from "axios";
+
+const basePath = "/api";
+
+export default {
+  props: ["to_delete", "to_delete_type"],
+  methods: {
+    deleteData() {
+      const path =
+        basePath + "/xss/" + this.to_delete + "/data/" + this.to_delete_type;
+
+      axios
+        .delete(path)
+        .then((response) => {
+          this.makeToast(response.data.detail, "success", response.data.status);
+          this.$emit("get-data");
+          this.$refs.deleteDataModal.hide();
+        })
+        .catch((error) => {
+          this.handleError(error);
+        });
+    },
+  },
+};
+</script>
+
+<style>
+</style>
