+++ client/src/components/index/viewData/ViewData.vue
@@ -5,8 +5,8 @@
     title="Captured data"
     hide-footer
     size="xl"
-    @show="getData"
-    @hide="cleanup"
+    @show="getData()"
+    @hide="cleanup()"
   >
     <b-row>
       <b-col offset-sm="8" sm="4">
@@ -205,24 +205,11 @@
       </b-table>
     </div>
 
-    <b-modal
-      ref="deleteDataModal"
-      id="delete-data-modal"
-      title="Are you sure?"
-      hide-footer
-    >
-      <b-form>
-        <b-button @click="deleteData" variant="outline-danger"
-          >Yes, delete this entry</b-button
-        >
-        <b-button
-          @click="$refs.deleteDataModal.hide()"
-          variant="outline-secondary"
-          >Cancel</b-button
-        >
-      </b-form>
-    </b-modal>
-
+    <DeleteData
+      :to_delete="to_delete"
+      :to_delete_type="to_delete_type"
+      @get-data="getData"
+    />
     <ViewDetails :xss_id="xss_detail_id" :client_id="client_id" />
   </b-modal>
 </template>
@@ -230,7 +217,8 @@
 <script>
 import axios from "axios";
 import VueJsonPretty from "vue-json-pretty";
-import ViewDetails from "./ViewDetails";
+import ViewDetails from "../shared/ViewDetails";
+import DeleteData from "./DeleteData";
 
 const basePath = "/api";
 
@@ -239,6 +227,7 @@ export default {
   components: {
     VueJsonPretty,
     ViewDetails,
+    DeleteData,
   },
   data() {
     return {
@@ -274,22 +263,7 @@ export default {
           this.data = response.data;
         })
         .catch((error) => {
-          this.$parent.handleAuthError(error);
-        });
-    },
-    deleteData() {
-      const path =
-        basePath + "/xss/" + this.to_delete + "/data/" + this.to_delete_type;
-
-      axios
-        .delete(path)
-        .then((response) => {
-          this.makeToast(response.data.detail, "success", response.data.status);
-          this.getData();
-          this.$refs.deleteDataModal.hide();
-        })
-        .catch((error) => {
-          this.$parent.handleAuthError(error);
+          this.handleError(error);
         });
     },
     getSpecificData(element_id, loot_type, row_index) {
@@ -308,26 +282,18 @@ export default {
           this.componentKey += 1;
         })
         .catch((error) => {
-          this.$parent.handleAuthError(error);
+          this.handleError(error);
         });
     },
     cleanSpecificData(element_id, loot_type, row_index) {
       this.data[loot_type][row_index][element_id] = "";
       this.componentKey += 1;
     },
-    makeToast(message, variant, title) {
-      this.$root.$bvToast.toast(message, {
-        title: title,
-        autoHideDelay: 5000,
-        appendToast: false,
-        variant: variant,
-      });
-    },
     cleanup() {
       this.data = {};
       this.to_delete = 0;
       this.to_delete_type = "";
-      this.$parent.getClients();
+      this.$emit("get-clients");
     },
   },
 };
