+++ client/src/components/login/Logo.vue
@@ -1,0 +1,19 @@
+<template>
+  <div>
+    <b-row>
+      <b-col>
+        <img height="200px" src="../../assets/images/logo.png" />
+      </b-col>
+    </b-row>
+    <b-row>
+      <b-col>
+        <img height="50px" src="../../assets/images/title.png" />
+      </b-col>
+    </b-row>
+  </div>
+</template>
+<script>
+export default {};
+</script>
+<style>
+</style>
