+++ client/src/components/shared/ChangePassword.vue
@@ -4,8 +4,8 @@
     id="change-password-modal"
     title="Change password"
     hide-footer
-    @hidden="cleanup"
-    :visible="$parent.show_password_modal"
+    @hidden="cleanup()"
+    :visible="show_password_modal"
   >
     <b-form>
       <b-form-group
@@ -15,7 +15,7 @@
         label-for="input-field-op"
       >
         <b-form-input
-          @keyup.enter="changePassword"
+          @keyup.enter="changePassword()"
           v-model="old_password"
           id="input-field-op"
           type="password"
@@ -30,7 +30,7 @@
         label-for="input-field-np"
       >
         <b-form-input
-          @keyup.enter="changePassword"
+          @keyup.enter="changePassword()"
           v-model="new_password1"
           id="input-field-np"
           type="password"
@@ -45,7 +45,7 @@
         label-for="input-field-np2"
       >
         <b-form-input
-          @keyup.enter="changePassword"
+          @keyup.enter="changePassword()"
           v-model="new_password2"
           id="input-field-np2"
           type="password"
@@ -53,8 +53,8 @@
         ></b-form-input>
       </b-form-group>
       <div class="text-right">
-        <b-button @click="changePassword" variant="outline-info">Save</b-button>
-        <b-button @click="cleanup" variant="outline-secondary">Cancel</b-button>
+        <b-button @click="changePassword()" variant="outline-info">Save</b-button>
+        <b-button @click="cleanup()" variant="outline-secondary">Cancel</b-button>
       </div>
     </b-form>
   </b-modal>
@@ -66,6 +66,7 @@ import axios from "axios";
 const basePath = "/api";
 
 export default {
+  props: ["show_password_modal"],
   data() {
     return {
       old_password: "",
@@ -90,16 +91,8 @@ export default {
           this.cleanup();
         })
         .catch((error) => {
-          this.$parent.handleAuthError(error);
+          this.handleError(error);
         });
-    },
-    makeToast(message, variant, title) {
-      this.$root.$bvToast.toast(message, {
-        title: title,
-        autoHideDelay: 5000,
-        appendToast: false,
-        variant: variant,
-      });
     },
     cleanup() {
       this.old_password = "";
