+++ vhost.conf
@@ -11,7 +11,7 @@ AddDefaultCharset UTF-8
 
         DocumentRoot /var/www/html/client/dist
 
-        <Location "/static/collector.min.js">
+        <Location "/static/scripts/collector.min.js">
             Header always set Cache-control "no-cache"
         </Location>
